fix - cambio de cantidad de likes de Int  a Long por un post que subieron con la cantidad de likes que superaba el valor de un Int.

+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/data/models/PostResponseDto.kt
@@ -8,7 +8,7 @@ data class PostResponseDto(
     @SerializedName("parent_id") val parentId: Int,
     val author: String,
     @SerializedName("avatar_url") val avatarUrl: String,
-    val likes: Int,
+    val likes: Long,
     val liked: Boolean,
     val date: String,
 )
